Add java doc
