+++ modules/eventlog/src/main/java/org/apache/ignite/internal/eventlog/api/IgniteEvents.java
@@ -15,11 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.eventlog.event;
+package org.apache.ignite.internal.eventlog.api;
 
 import java.util.Arrays;
-import org.apache.ignite.internal.eventlog.api.Event;
-import org.apache.ignite.internal.eventlog.api.EventFactory;
+import org.apache.ignite.internal.eventlog.event.EventBuilder;
+import org.apache.ignite.internal.eventlog.event.EventTypeRegistry;
+import org.apache.ignite.internal.eventlog.event.EventUser;
 
 /**
  * The main class for creating all Ignite events.
@@ -35,6 +36,8 @@ public final class IgniteEvents implements EventFactory {
     public static final IgniteEvents CONNECTION_CLOSED = new IgniteEvents(IgniteEventType.CONNECTION_CLOSED.name());
 
     static {
+        // Without the following line, the IgniteEventType enum will not be registered in the EventTypeRegistry
+        // and the EventTypeRegistry will not be able to validate the event types.
         Arrays.stream(IgniteEventType.values()).forEach(type -> EventTypeRegistry.register(type.name()));
     }
