+++ modules/eventlog/src/main/java/org/apache/ignite/internal/eventlog/api/Sink.java
@@ -17,11 +17,24 @@
 
 package org.apache.ignite.internal.eventlog.api;
 
-import java.util.Set;
-import org.apache.ignite.internal.eventlog.sink.Sink;
-
-public interface SinkRegistry {
-    Sink getByName(String name);
-
-    Set<Sink> findAllByChannel(String channel);
+/**
+ * The endpoint for the event log framework. This is the last step in the event log pipeline. It can be a log file, a webhook, a Kafka
+ * topic, or whatever we develop.
+ *
+ * <p>The contract of the only method is the following:
+ *
+ * <p>IT DOES NOT GUARANTEE THAT THE EVENT IS WRITTEN TO THE FINAL DESTINATION.
+ * For example, if the sink as a log file, the method does not guarantee that the event is written to the file. Because the logging
+ * framework can be asynchronous.
+ *
+ * <p>IT DOES GUARANTEE THAT THE EVENT IS SENT TO THE SINK.
+ * For example, if the sink is a Kafka topic, the method guarantees that the event is sent to the topic.
+ */
+public interface Sink {
+    /**
+     * Writes the event to the sink.
+     *
+     * @param event The event to write.
+     */
+    void write(Event event);
 }
