+++ modules/eventlog/src/main/java/org/apache/ignite/internal/eventlog/config/schema/ChannelConfigurationSchema.java
@@ -20,16 +20,21 @@ package org.apache.ignite.internal.eventlog.config.schema;
 import org.apache.ignite.configuration.annotation.Config;
 import org.apache.ignite.configuration.annotation.InjectedName;
 import org.apache.ignite.configuration.annotation.Value;
+import org.apache.ignite.internal.eventlog.api.IgniteEventType;
 
+/** Channel configuration schema. */
 @Config
 public class ChannelConfigurationSchema {
 
+    /** Channel name. This name is used to reference from sink configuration. */
     @InjectedName
     public String name;
 
+    /** Channel enabled flag. */
     @Value(hasDefault = true)
     public boolean enabled = true;
 
+    /** Event types that this channel should handle. Should be the types from {@link IgniteEventType}. */
     @Value(hasDefault = true)
     public String[] events = {};
 }
