+++ modules/eventlog/src/main/java/org/apache/ignite/internal/eventlog/impl/SinkRegistry.java
@@ -15,13 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.eventlog.event;
+package org.apache.ignite.internal.eventlog.impl;
 
-/**
- * Defines a subset of event types that can be created in the system. Note, the event type is a string that is unique
- * within the system. The event type is used to filter the events in the event log.
- */
-public enum IgniteEventType {
-    USER_AUTHENTICATED,
-    CONNECTION_CLOSED
+import java.util.Set;
+import org.apache.ignite.internal.eventlog.api.Sink;
+
+interface SinkRegistry {
+    Sink getByName(String name);
+
+    Set<Sink> findAllByChannel(String channel);
 }
